CredEquate: Fix circular dependencies
